requirements.txt
requests>=2.28
pandas>=1.5
pyarrow>=10.0   # optional (for parquet output)
python-dateutil
tqdm

